requirements.txt
# Medical Prescription Firewall - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
FastAPI==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.0
openpyxl==3.11.0
numpy==1.24.0

# Google Sheets Integration
gspread==5.11.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.25.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Logging & Monitoring
loguru==0.7.2

# Performance
gunicorn==21.2.0

# Optional: Streamlit Dashboard
streamlit==1.29.0

# Optional: Docker
docker==7.0.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1

# CORS & Middleware
requests==2.31.0

# API Documentation
swagger-ui==0.1.2
